requirements.txt
# Python dependencies for Jarvis Home Automation

# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# System monitoring (optional - will be used in future for Windows/Mac collectors)
psutil>=5.9.0

# Future dependencies (uncomment as needed)
# requests>=2.31.0  # For API interactions
# aiohttp>=3.9.0  # For async HTTP requests
# pyshelly>=0.1.0  # For Shelly integration (if available)
# pytapo>=0.3.0  # For Tapo integration
# alexapy>=1.27.0  # For Alexa integration

# Development dependencies
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# pylint>=2.17.0
# mypy>=1.5.0
